fix: monitoring scan mailbox not working
IDLE acts on the selected mailbox.
The hamMailbox can be the last selected one, causing that the wrong
mailbox is selected.
SELECT the scan mailbox explicitly before starting IDLE.
Also remove subscribing to scan mailbox, this is not needed.

If SELECT returns that the mailbox has new messages, IDLE is not
started, the new messages are processed immediately.