imap: add scan result to mails
The rspamd scan symbols with their value are now added to scanned
emails.
Emails in the ScanMailbox are now downloaded and stored in the directory
that can be configured via the TempDir directive in the configuration
file.
Afterwards they are passed to rspamd for scanning.
Rspamd returns the total score and the invididual score symbols.
A local copy of the downloaded email is created, where the symbols are
added as additional e-mail headers, prefixed with "X-rspamd-iscan-".
All symbols with a 0 score are omitted.
The original email is then moved to a backup mailbox, that can be
configured via the new BackupMailbox directive in the config file.
Afterwards the modified e-mail is uploaded to the INBOX mailbox.

The backup mailbox exist, to prevent that e-mails are lost because of
bugs this new feature. A configuration option can be added later to
disable backing up the original e-mails.
Currently the user has to delete old emails from the backup mailbox
manually.

Also:
- mailboxes are not opened read-only anymore when messages are moved
  afterwards, when fetching messages of those the Peek flag is now set
  correctly to prevent that their status changed to read
- The stateFile is removed. This means that the ScanMailbox and
  InboxMailbox can't be the same mailbox anymore.
- When errors happening while scanning e-mails, the erroneous ones are skipped and scanning continues.
  The operations returns all collected errors at the end.
  Previously the scan operation aborted as soon as an error happened,
  not continuing to scan other available emails.