split imap package
Split the "imap" package into "imapclt" and "iscan". "imapclt" provides
the required IMAP functionality with a simpler interface by wrapping
go-imap.
"iscan" provides the connector between the IMAP-server and rspamd.

The allows to implement a dry-run mode, that does not
modify the content on the IMAP-server, in a follow-up.

Also:
- Testcases are added using emersion/go-imap/v2/imapserver,
- iscan.Stop() now waits until a parallel running Start() go-routine has
  terminated,
- Config validation is moved to the iscan package