iscan: learn spam and ham every 30min
Spam and Ham are now learned every 30min.
Before Spam and Ham messages were learned when a 30min timer expired and
the last learning happened more than 30min ago or a new message in the
Scanbox was detected and learning happened more than 30min ago.

This could cause that in some situation ham and spam was learned less
often that every 30min. This isn't an issue because learning ham and
spam isn't a a critical task and can be delayed.

To make it deterministic for testing the behaviour in testcases, it is
now learned strictly every c.learnInterval.