fix: close is called for former iscan.Clients instances on retry
When a retryable error occurs in main(), the signal handler that calls
close on iscan.Client is kept installed.

This means the iscan.Client instances must also be kept in memory. This
is a memory leak.

It is fixed by refactoring main() and removing all installed signal
handlers when a client returns an error, before a new one is created.

It would be good to split creating a Client and establishing the
connection into 2 methods in a follow-up, to prevent recreating the
iscan.Client unnecessarily on every retry.