add --dry-run command line flag
When --dry-run is passed, a special IMAP client is used that only
pretends to Move and Upload E-Mails.

This can be used for testing the rspamd-iscan functionality, without the
risk of altering IMAP mailboxes.

--dry-run also enables --once, this is required to prevent that in
monitoring mode the same messages will be processed over and over again.
--dry-run could be supported in monitoring mode if the already processed
messages would be tracked. It's not worth to add the functionality only
for the dry-run mode though.