imap: allow logging raw IMAP data for debugging
To be able to debug issues with the imap-client the new config option
LogIMAPData is added.
When enabled, the raw exchange of the IMAP data with the server is
logged with debug log level.

This must be enabled with care, because the logged data contains
sensitive information, like the the IMAP-Server credentials and message
data.